refactor: 유틸리티 클래스의 패키지 변경
