+++ src/main/java/nextstep/subway/util/FareCalculator.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion;
+package nextstep.subway.util;
 
 public class FareCalculator {
