+++ src/test/java/nextstep/subway/unit/FareCalculatorTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import nextstep.subway.applicaion.FareCalculator;
+import nextstep.subway.util.FareCalculator;
 
 class FareCalculatorTest {
